refactor(infra): extract shared privileged converge from apply + db-exposure
runApply and convergePublicEndpoint were the same bootstrap-key converge
ritual (passphrase, key prompts, lock, orphan adoption, sync-rollout-config,
pulumi-up retry loop) maintained twice with cosmetic drift. One
runPrivilegedConverge now owns it, taking the better half of each parent:
apply's state-identity override and guarded finally lock release, exposure's
--config-file prepare hook. pulumiConfigSet/Rm move to lib/stack/pulumi-up
(plain stack-config ops), and the two secret-output readers collapse into one.
